requirements.txt
google-genai
python-dotenv
langchain-community
faiss-cpu
langchain-cohere
pymupdf
streamlit
nest_asyncio